+++ package.json
@@ -21,6 +21,7 @@
     "postcss": "8.4.30",
     "react": "18.2.0",
     "react-calendar": "^4.6.0",
+    "react-cookie": "^6.1.1",
     "react-dom": "18.2.0",
     "react-icons": "^4.11.0",
     "react-redux": "^8.1.2",
